feat: Add a service for the Rental entity
Add the table name "rentals". Create a method "bookItem" in the "RentalService"
that converts "RentalDto" object to "Rental" object and calls the saving method.

+++ src/main/java/mb/equipme_rent_service/web/model/RentalDto.java
@@ -18,7 +18,7 @@ public class RentalDto {
     private String upc;
 
     @JsonProperty("rentalPrice")
-    private Long rentalPrice;
+    private Double rentalPrice;
 
     @JsonProperty("rentalStartDate")
     private LocalDate rentalStartDate;
